refactor(apps): todo apps
